added sub_services to the namespace of services in routes.rb

+++ db/migrate/20090312083036_create_feedsettings.rb
@@ -1,8 +1,8 @@
 class CreateFeedsettings < ActiveRecord::Migration
   def self.up
     create_table :feedsettings do |t|
-      t.decimal :user_id
-      t.decimal :service_id
+      t.integer :user_id
+      t.integer :service_id
       t.string :parameter
 
       t.timestamps
